refactor: dedup helpers, split tests, fix get_callers params, add .env.example
