+++ .env.example
@@ -1,0 +1,51 @@
+# CodeLens MCP Plugin — Environment Variables
+# Copy to .env and customize for your deployment.
+# Both CODELENS_* and SYMBIOTE_* prefixes are accepted (SYMBIOTE wins).
+
+# ── Transport / Daemon ──
+# CODELENS_DAEMON_MODE=
+# CODELENS_EXECUTABLE_PATH_OVERRIDE=
+
+# ── Auth / Security ──
+# CODELENS_AUTH_MODE=off         # off | jwt | mTLS
+# CODELENS_AUTH_ISSUER=
+# CODELENS_AUTH_AUDIENCE=
+# CODELENS_AUTH_JWKS_URL=
+# CODELENS_AUTH_SCOPE=
+# CODELENS_PRINCIPAL=
+# CODELENS_RATE_LIMIT=
+
+# ── Embedding / Semantic Search ──
+# CODELENS_EMBED_BATCH_SIZE=32
+# CODELENS_EMBED_DOCSTRINGS=1
+# CODELENS_EMBED_COREML_COMPUTE_UNITS=ALL
+# CODELENS_EMBED_COREML_MODEL_FORMAT=MLprogram
+# CODELENS_EMBED_HINT_AUTO=1
+# CODELENS_EMBED_HINT_AUTO_LANG=rust,python,typescript
+# CODELENS_EMBED_HINT_CHARS=8000
+# CODELENS_EMBED_HINT_LINES=200
+# CODELENS_EMBED_HINT_INCLUDE_COMMENTS=1
+# CODELENS_EMBED_HINT_INCLUDE_API_CALLS=1
+# CODELENS_EMBED_HINT_STRICT_COMMENTS=0
+# CODELENS_EMBED_HINT_STRICT_LITERALS=0
+
+# ── External Tool Adapters ──
+# CODELENS_JETBRAINS_ADAPTER_CMD=
+# CODELENS_ROSLYN_ADAPTER_CMD=
+# CODELENS_LSP_PATH_EXTRA=
+# CODELENS_GENERIC_BRIDGES_OFF=0
+# CODELENS_PROJECT_BRIDGES_ON=0
+
+# ── Telemetry / Audit ──
+# CODELENS_TELEMETRY_ENABLED=0
+# CODELENS_TELEMETRY_PATH=
+# CODELENS_AUDIT_RETENTION_DAYS=30
+
+# ── Preflight / Caching ──
+# CODELENS_PREFLIGHT_TTL_SECS=300
+
+# ── Response Format ──
+# CODELENS_VERBOSE_TEXT=0        # 1 = full data in text payload
+
+# ── Effort / Workflow ──
+# CODELENS_EFFORT_LEVEL=balanced # minimal | balanced | thorough
