+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "crates/codelens-engine",
     "crates/codelens-mcp",
-    "crates/codelens-tui",
 ]
 default-members = [
     "crates/codelens-engine",
