+++ crates/codelens-mcp/src/artifact_store.rs
@@ -269,6 +269,7 @@ impl AnalysisArtifactStore {
         Some(artifact)
     }
 
+    #[allow(dead_code)] // backward-compatibility wrapper for external callers
     pub fn find_reusable(
         &self,
         tool_name: &str,
@@ -276,6 +277,21 @@ impl AnalysisArtifactStore {
         surface_label: &str,
         project_scope: Option<&str>,
     ) -> Option<AnalysisArtifact> {
+        self.find_reusable_tiered(tool_name, cache_key, surface_label, project_scope)
+            .map(|(artifact, _)| artifact)
+    }
+
+    /// Tiered cache lookup:
+    /// - L1 (Exact):  tool_name + cache_key + surface + scope match
+    /// - L2 (Warm):   tool_name + surface + scope match (any cache_key)
+    /// - L3 (Cold):   same scope, any tool, most recent within TTL
+    pub fn find_reusable_tiered(
+        &self,
+        tool_name: &str,
+        cache_key: &str,
+        surface_label: &str,
+        project_scope: Option<&str>,
+    ) -> Option<(AnalysisArtifact, crate::runtime_types::CacheHitTier)> {
         self.prune(crate::util::now_ms());
         for id in self.list_ids_on_disk() {
             let _ = self.get(&id, project_scope);
@@ -289,14 +305,41 @@ impl AnalysisArtifactStore {
             .cloned()
             .collect::<Vec<_>>();
         let arts = self.artifacts.lock().unwrap_or_else(|p| p.into_inner());
-        order.into_iter().find_map(|id| {
-            let a = arts.get(&id)?;
+
+        // L1: exact match
+        if let Some(artifact) = order.iter().find_map(|id| {
+            let a = arts.get(id)?;
             (a.tool_name == tool_name
                 && a.surface == surface_label
                 && matches_scope(a.project_scope.as_deref(), project_scope)
                 && a.cache_key.as_deref() == Some(cache_key))
             .then(|| a.clone())
-        })
+        }) {
+            return Some((artifact, crate::runtime_types::CacheHitTier::Exact));
+        }
+
+        // L2: warm match — same tool + surface + scope, stored artifact has no cache_key (generic)
+        if let Some(artifact) = order.iter().find_map(|id| {
+            let a = arts.get(id)?;
+            (a.tool_name == tool_name
+                && a.surface == surface_label
+                && matches_scope(a.project_scope.as_deref(), project_scope)
+                && a.cache_key.is_none())
+            .then(|| a.clone())
+        }) {
+            return Some((artifact, crate::runtime_types::CacheHitTier::Warm));
+        }
+
+        // L3: cold match — same scope, stored artifact has no cache_key (generic), most recent
+        if let Some(artifact) = order.iter().find_map(|id| {
+            let a = arts.get(id)?;
+            (matches_scope(a.project_scope.as_deref(), project_scope) && a.cache_key.is_none())
+                .then(|| a.clone())
+        }) {
+            return Some((artifact, crate::runtime_types::CacheHitTier::Cold));
+        }
+
+        None
     }
 
     pub fn list_summaries(&self, project_scope: Option<&str>) -> Vec<AnalysisSummary> {
@@ -366,5 +409,184 @@ fn matches_scope(artifact_scope: Option<&str>, current_scope: Option<&str>) -> b
         (Some(a), Some(c)) => a == c,
         (None, _) => true,
         (_, None) => true,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::runtime_types::{AnalysisArtifact, AnalysisReadiness, CacheHitTier};
+    use std::collections::BTreeMap;
+
+    fn make_test_artifact(
+        id: &str,
+        tool_name: &str,
+        surface: &str,
+        project_scope: Option<&str>,
+        cache_key: Option<&str>,
+        created_at_ms: u64,
+    ) -> AnalysisArtifact {
+        AnalysisArtifact {
+            id: id.to_owned(),
+            tool_name: tool_name.to_owned(),
+            surface: surface.to_owned(),
+            project_scope: project_scope.map(|s| s.to_owned()),
+            cache_key: cache_key.map(|s| s.to_owned()),
+            summary: "test".to_owned(),
+            top_findings: vec!["finding".to_owned()],
+            risk_level: "medium".to_owned(),
+            confidence: 0.5,
+            next_actions: vec!["act".to_owned()],
+            blockers: vec![],
+            readiness: AnalysisReadiness::default(),
+            verifier_checks: vec![],
+            available_sections: vec!["summary".to_owned()],
+            created_at_ms,
+        }
+    }
+
+    fn make_store(tmp: &std::path::Path) -> AnalysisArtifactStore {
+        let dir = tmp.join("artifacts");
+        std::fs::create_dir_all(&dir).unwrap();
+        AnalysisArtifactStore::new(dir)
+    }
+
+    fn store_artifact(store: &AnalysisArtifactStore, artifact: AnalysisArtifact) {
+        store
+            .store(
+                &artifact.tool_name,
+                &artifact.surface,
+                artifact.project_scope.clone().unwrap_or_default(),
+                artifact.cache_key.clone(),
+                artifact.summary.clone(),
+                artifact.top_findings.clone(),
+                artifact.risk_level.clone(),
+                artifact.confidence,
+                artifact.next_actions.clone(),
+                artifact.blockers.clone(),
+                artifact.readiness.clone(),
+                artifact.verifier_checks.clone(),
+                std::collections::BTreeMap::from([("summary".to_owned(), serde_json::json!({}))]),
+            )
+            .unwrap();
+    }
+
+    #[test]
+    fn tiered_exact_hit() {
+        let tmp = std::env::temp_dir().join(format!("codelens-test-{}", std::process::id()));
+        let _ = std::fs::remove_dir_all(&tmp);
+        let store = make_store(&tmp);
+        let artifact = make_test_artifact(
+            "a1",
+            "impact_report",
+            "full",
+            Some("/proj"),
+            Some("key1"),
+            crate::util::now_ms(),
+        );
+        store_artifact(&store, artifact);
+        let (found, tier) = store
+            .find_reusable_tiered("impact_report", "key1", "full", Some("/proj"))
+            .unwrap();
+        assert_eq!(found.tool_name, "impact_report");
+        assert_eq!(tier, CacheHitTier::Exact);
+        let _ = std::fs::remove_dir_all(&tmp);
+    }
+
+    #[test]
+    fn tiered_warm_hit_same_tool_no_cache_key() {
+        let tmp = std::env::temp_dir().join(format!("codelens-test-{}", std::process::id()));
+        let _ = std::fs::remove_dir_all(&tmp);
+        let store = make_store(&tmp);
+        let artifact = make_test_artifact(
+            "a1",
+            "impact_report",
+            "full",
+            Some("/proj"),
+            None, // generic analysis
+            crate::util::now_ms(),
+        );
+        store_artifact(&store, artifact);
+        let (found, tier) = store
+            .find_reusable_tiered("impact_report", "new-key", "full", Some("/proj"))
+            .unwrap();
+        assert_eq!(found.tool_name, "impact_report");
+        assert_eq!(tier, CacheHitTier::Warm);
+        let _ = std::fs::remove_dir_all(&tmp);
+    }
+
+    #[test]
+    fn tiered_cold_hit_different_tool_same_scope_generic() {
+        let tmp = std::env::temp_dir().join(format!("codelens-test-{}", std::process::id()));
+        let _ = std::fs::remove_dir_all(&tmp);
+        let store = make_store(&tmp);
+        let artifact = make_test_artifact(
+            "a1",
+            "impact_report",
+            "full",
+            Some("/proj"),
+            None, // generic analysis
+            crate::util::now_ms(),
+        );
+        store_artifact(&store, artifact);
+        let (found, tier) = store
+            .find_reusable_tiered("change_request", "other-key", "full", Some("/proj"))
+            .unwrap();
+        assert_eq!(found.tool_name, "impact_report");
+        assert_eq!(tier, CacheHitTier::Cold);
+        let _ = std::fs::remove_dir_all(&tmp);
+    }
+
+    #[test]
+    fn tiered_miss_different_scope() {
+        let tmp = std::env::temp_dir().join(format!("codelens-test-{}", std::process::id()));
+        let _ = std::fs::remove_dir_all(&tmp);
+        let store = make_store(&tmp);
+        let artifact = make_test_artifact(
+            "a1",
+            "impact_report",
+            "full",
+            Some("/other"),
+            Some("key1"),
+            crate::util::now_ms(),
+        );
+        store_artifact(&store, artifact);
+        assert!(
+            store
+                .find_reusable_tiered("impact_report", "key1", "full", Some("/proj"))
+                .is_none()
+        );
+        let _ = std::fs::remove_dir_all(&tmp);
+    }
+
+    #[test]
+    fn tiered_exact_preferred_over_warm() {
+        let tmp = std::env::temp_dir().join(format!("codelens-test-{}", std::process::id()));
+        let _ = std::fs::remove_dir_all(&tmp);
+        let store = make_store(&tmp);
+        let warm = make_test_artifact(
+            "warm",
+            "impact_report",
+            "full",
+            Some("/proj"),
+            None, // generic → eligible for warm
+            crate::util::now_ms() - 1000,
+        );
+        let exact = make_test_artifact(
+            "exact",
+            "impact_report",
+            "full",
+            Some("/proj"),
+            Some("key1"),
+            crate::util::now_ms(),
+        );
+        store_artifact(&store, warm);
+        store_artifact(&store, exact);
+        let (found, tier) = store
+            .find_reusable_tiered("impact_report", "key1", "full", Some("/proj"))
+            .unwrap();
+        assert_eq!(found.cache_key.as_deref(), Some("key1"));
+        assert_eq!(tier, CacheHitTier::Exact);
+        let _ = std::fs::remove_dir_all(&tmp);
     }
 }
