+++ crates/codelens-mcp/src/integration_tests/protocol_security.rs
@@ -1,0 +1,91 @@
+use super::*;
+
+// ── Security, daemon mode, and surface gating ─────────────────────────────────────
+
+#[test]
+fn tool_call_result_meta_exposes_preferred_executor() {
+    let project = project_root();
+    let state = crate::AppState::new(project, crate::tool_defs::ToolPreset::Full);
+
+    let response = handle_request(
+        &state,
+        crate::protocol::JsonRpcRequest {
+            jsonrpc: "2.0".to_owned(),
+            id: Some(json!(1)),
+            method: "tools/call".to_owned(),
+            params: Some(json!({
+                "name": "create_text_file",
+                "arguments": {
+                    "_session_id": default_session_id(&state),
+                    "relative_path": "new_file.txt",
+                    "content": "hello\n"
+                }
+            })),
+        },
+    )
+    .expect("tools/call should return a response");
+
+    let value = serde_json::to_value(&response).expect("serialize");
+    assert_eq!(
+        value["result"]["_meta"]["codelens/preferredExecutor"],
+        json!("codex-builder")
+    );
+}
+
+#[test]
+fn read_only_daemon_rejects_mutation_even_with_mutating_profile() {
+    let project = project_root();
+    let state = crate::AppState::new(project, crate::tool_defs::ToolPreset::Full);
+    state.set_surface(crate::tool_defs::ToolSurface::Profile(
+        crate::tool_defs::ToolProfile::RefactorFull,
+    ));
+    state.configure_daemon_mode(crate::state::RuntimeDaemonMode::ReadOnly);
+
+    let payload = call_tool(
+        &state,
+        "create_text_file",
+        json!({"relative_path": "blocked.txt", "content": "nope"}),
+    );
+    assert_eq!(payload["success"], json!(false));
+    assert!(
+        payload["error"]
+            .as_str()
+            .unwrap_or("")
+            .contains("blocked by daemon mode")
+    );
+}
+
+#[test]
+fn hidden_tools_are_blocked_at_call_time() {
+    let project = project_root();
+    let state = crate::AppState::new(project, crate::tool_defs::ToolPreset::Full);
+    let _ = call_tool(
+        &state,
+        "set_profile",
+        json!({"profile": "planner-readonly"}),
+    );
+
+    let payload = call_tool(
+        &state,
+        "create_text_file",
+        json!({"relative_path": "blocked.txt", "content": "nope"}),
+    );
+    assert_eq!(payload["success"], json!(false));
+    assert!(
+        payload["error"]
+            .as_str()
+            .unwrap_or("")
+            .contains("not available in active surface")
+    );
+}
+
+#[test]
+fn read_only_surface_marks_content_mutations_for_blocking() {
+    assert!(crate::tool_defs::is_read_only_surface(
+        crate::tool_defs::ToolSurface::Profile(crate::tool_defs::ToolProfile::PlannerReadonly),
+    ));
+    assert!(crate::tool_defs::is_content_mutation_tool(
+        "create_text_file"
+    ));
+    assert!(!crate::tool_defs::is_content_mutation_tool("set_profile"));
+}
