+++ crates/codelens-mcp/src/tool_defs/presets.rs
@@ -228,19 +228,13 @@ impl SurfaceCompilerInput {
     }
 }
 
-fn push_unique(items: &mut Vec<&'static str>, value: &'static str) {
-    if !items.contains(&value) {
-        items.push(value);
-    }
-}
-
 fn push_tool_if_in_surface(
     items: &mut Vec<&'static str>,
     surface: ToolSurface,
     tool: &'static str,
 ) {
     if is_tool_in_surface(tool, surface) {
-        push_unique(items, tool);
+        crate::util::push_unique(items, tool);
     }
 }
 
@@ -251,6 +245,23 @@ fn push_entrypoint_if_in_surface(
 ) {
     if tool != "prepare_harness_session" {
         push_tool_if_in_surface(items, surface, tool);
+    }
+}
+
+/// Push a list of tools into both entrypoints and emphasized_tools,
+/// but only if each tool is included in the given surface.
+fn push_surface_tools(plan: &mut SurfaceOverlayPlan, surface: ToolSurface, tools: &[&'static str]) {
+    for tool in tools {
+        push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
+        push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
+    }
+}
+
+/// Push a list of tools into avoid_tools, but only if each tool is
+/// included in the given surface.
+fn push_avoid_tools(plan: &mut SurfaceOverlayPlan, surface: ToolSurface, tools: &[&'static str]) {
+    for tool in tools {
+        push_tool_if_in_surface(&mut plan.avoid_tools, surface, tool);
     }
 }
 
@@ -273,105 +284,112 @@ pub(crate) fn compile_surface_overlay(
         match host_context {
             HostContext::ClaudeCode => {
                 plan.preferred_executor_bias = Some("claude");
-                for tool in [
-                    "prepare_harness_session",
-                    "analyze_change_request",
-                    "review_changes",
-                    "impact_report",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "analyze_change_request",
+                        "review_changes",
+                        "impact_report",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Claude Code hosts should stay in workflow/report lanes first and only cross into builder-heavy execution when the executor hint or delegate scaffold says so.",
                 );
             }
             HostContext::Codex => {
                 plan.preferred_executor_bias = Some("codex-builder");
-                for tool in [
-                    "prepare_harness_session",
-                    "explore_codebase",
-                    "trace_request_path",
-                    "plan_safe_refactor",
-                    "verify_change_readiness",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "explore_codebase",
+                        "trace_request_path",
+                        "plan_safe_refactor",
+                        "verify_change_readiness",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Codex hosts should bias toward compact bootstrap and execution-oriented follow-up tools instead of broad planner chatter.",
                 );
             }
             HostContext::Cursor => {
-                for tool in [
-                    "prepare_harness_session",
-                    "explore_codebase",
-                    "trace_request_path",
-                    "review_changes",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "explore_codebase",
+                        "trace_request_path",
+                        "review_changes",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Cursor sessions should keep the initial MCP surface compact and lean on workflow-level tools before expanding primitives.",
                 );
             }
             HostContext::Cline => {
-                for tool in [
-                    "prepare_harness_session",
-                    "review_changes",
-                    "get_file_diagnostics",
-                    "verify_change_readiness",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "review_changes",
+                        "get_file_diagnostics",
+                        "verify_change_readiness",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Cline sessions benefit from explicit review and diagnostics checkpoints before mutation-heavy chains.",
                 );
             }
             HostContext::Windsurf => {
-                for tool in [
-                    "prepare_harness_session",
-                    "explore_codebase",
-                    "trace_request_path",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "explore_codebase",
+                        "trace_request_path",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Windsurf hosts have a tighter MCP budget, so keep the surface bounded and prefer high-signal workflow entrypoints.",
                 );
             }
             HostContext::VsCode | HostContext::JetBrains => {
-                for tool in [
-                    "prepare_harness_session",
-                    "explore_codebase",
-                    "review_changes",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "explore_codebase",
+                        "review_changes",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "IDE hosts should use CodeLens for bootstrap, review, and bounded context retrieval rather than mirroring the full editor-native toolchain.",
                 );
             }
             HostContext::ApiAgent => {
-                for tool in [
-                    "prepare_harness_session",
-                    "start_analysis_job",
-                    "get_analysis_section",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "start_analysis_job",
+                        "get_analysis_section",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "API agents should prefer compact bootstrap and durable analysis handles over long in-band transcripts.",
                 );
@@ -382,121 +400,133 @@ pub(crate) fn compile_surface_overlay(
     if let Some(task_overlay) = task_overlay {
         match task_overlay {
             TaskOverlay::Planning => {
-                for tool in [
-                    "prepare_harness_session",
-                    "explore_codebase",
-                    "analyze_change_request",
-                    "review_architecture",
-                    "impact_report",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                for tool in [
-                    "rename_symbol",
-                    "replace_symbol_body",
-                    "insert_content",
-                    "replace",
-                    "add_import",
-                ] {
-                    push_tool_if_in_surface(&mut plan.avoid_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "explore_codebase",
+                        "analyze_change_request",
+                        "review_architecture",
+                        "impact_report",
+                    ],
+                );
+                push_avoid_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "rename_symbol",
+                        "replace_symbol_body",
+                        "insert_content",
+                        "replace",
+                        "add_import",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Planning overlay keeps the session in analyze/review lanes until the change boundary and acceptance checks are explicit.",
                 );
             }
             TaskOverlay::Editing => {
-                for tool in [
-                    "prepare_harness_session",
-                    "trace_request_path",
-                    "plan_safe_refactor",
-                    "verify_change_readiness",
-                    "get_file_diagnostics",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                for tool in [
-                    "rename_symbol",
-                    "replace_symbol_body",
-                    "insert_content",
-                    "replace",
-                    "add_import",
-                ] {
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "trace_request_path",
+                        "plan_safe_refactor",
+                        "verify_change_readiness",
+                        "get_file_diagnostics",
+                    ],
+                );
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "rename_symbol",
+                        "replace_symbol_body",
+                        "insert_content",
+                        "replace",
+                        "add_import",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Editing overlay expects trace -> preflight -> mutation -> diagnostics instead of repeated low-level search loops.",
                 );
             }
             TaskOverlay::Review => {
-                for tool in [
-                    "prepare_harness_session",
-                    "review_changes",
-                    "impact_report",
-                    "diff_aware_references",
-                    "audit_planner_session",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                for tool in [
-                    "rename_symbol",
-                    "replace_symbol_body",
-                    "insert_content",
-                    "replace",
-                ] {
-                    push_tool_if_in_surface(&mut plan.avoid_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "review_changes",
+                        "impact_report",
+                        "diff_aware_references",
+                        "audit_planner_session",
+                    ],
+                );
+                push_avoid_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "rename_symbol",
+                        "replace_symbol_body",
+                        "insert_content",
+                        "replace",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Review overlay keeps the session in evidence and audit lanes; mutation tools remain secondary unless the task explicitly escalates.",
                 );
             }
             TaskOverlay::Onboarding => {
-                for tool in [
-                    "prepare_harness_session",
-                    "onboard_project",
-                    "explore_codebase",
-                    "review_architecture",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "onboard_project",
+                        "explore_codebase",
+                        "review_architecture",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Onboarding overlay should favor durable project summaries and architectural overviews before narrow task routing.",
                 );
             }
             TaskOverlay::BatchAnalysis => {
-                for tool in [
-                    "prepare_harness_session",
-                    "start_analysis_job",
-                    "get_analysis_job",
-                    "get_analysis_section",
-                    "module_boundary_report",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "start_analysis_job",
+                        "get_analysis_job",
+                        "get_analysis_section",
+                        "module_boundary_report",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Batch-analysis overlay should move heavy work onto durable analysis jobs instead of keeping the host in a long synchronous loop.",
                 );
             }
             TaskOverlay::Interactive => {
-                for tool in [
-                    "prepare_harness_session",
-                    "explore_codebase",
-                    "find_symbol",
-                    "get_ranked_context",
-                ] {
-                    push_entrypoint_if_in_surface(&mut plan.preferred_entrypoints, surface, tool);
-                    push_tool_if_in_surface(&mut plan.emphasized_tools, surface, tool);
-                }
-                push_unique(
+                push_surface_tools(
+                    &mut plan,
+                    surface,
+                    &[
+                        "prepare_harness_session",
+                        "explore_codebase",
+                        "find_symbol",
+                        "get_ranked_context",
+                    ],
+                );
+                crate::util::push_unique(
                     &mut plan.routing_notes,
                     "Interactive overlay should keep bootstrap light and bias toward retrieval tools that answer the next question quickly.",
                 );
@@ -566,7 +596,6 @@ pub(crate) const BALANCED_EXCLUDES: &[&str] = &[
     "get_tool_metrics",
     "audit_builder_session",
     "audit_planner_session",
-    "export_session_markdown",
     "export_session_markdown",
     "summarize_changes",
     "summarize_file",
