+++ crates/codelens-mcp/src/tools/reports/impact_reports/mod.rs
@@ -6,7 +6,7 @@ mod refactor;
 
 use helpers::{
     build_dead_code_semantic_query, build_module_semantic_query, file_name, impact_entry_file,
-    insert_semantic_status, mermaid_escape_label, parent_dir, push_unique, semantic_degraded_note,
+    insert_semantic_status, mermaid_escape_label, parent_dir, semantic_degraded_note,
 };
 #[cfg(test)]
 use mermaid::render_module_mermaid;
